requirements.txt
Django == 3.1.2
djangorestframework == 3.12.1
djangorestframework-simplejwt == 4.4.0
wheel == 0.35.1
Pillow == 7.2.0
python-dotenv == 0.14.0